Implement basic SDK for convai agents
Early prototype, subject to change based on user feedback.

Takes care of the websocket session and message handling, exposing a
simplified audio interface to the client that can be hooked up to
the appropriate audio inputs / outputs based on the usecase.

Also implements a basic speaker/microphone interface, via optional
dependency on pyaudio.